+++ packages/core/src/index.ts
@@ -1,10 +1,13 @@
-export { default as ConfirmPrompt } from './prompts/confirm';
-export { default as GroupMultiSelectPrompt } from './prompts/group-multiselect';
-export { default as MultiSelectPrompt } from './prompts/multi-select';
-export { default as PasswordPrompt } from './prompts/password';
-export { default as Prompt } from './prompts/prompt';
-export { default as SelectPrompt } from './prompts/select';
-export { default as SelectKeyPrompt } from './prompts/select-key';
-export { default as TextPrompt } from './prompts/text';
-export type { ClackState as State } from './types';
-export { block, isCancel, setGlobalAliases } from './utils';
+export type { ClackState as State } from './types.js';
+export type { ClackSettings } from './utils/settings.js';
+
+export { default as ConfirmPrompt } from './prompts/confirm.js';
+export { default as GroupMultiSelectPrompt } from './prompts/group-multiselect.js';
+export { default as MultiSelectPrompt } from './prompts/multi-select.js';
+export { default as PasswordPrompt } from './prompts/password.js';
+export { default as Prompt } from './prompts/prompt.js';
+export { default as SelectPrompt } from './prompts/select.js';
+export { default as SelectKeyPrompt } from './prompts/select-key.js';
+export { default as TextPrompt } from './prompts/text.js';
+export { block, isCancel } from './utils/index.js';
+export { updateSettings, settings } from './utils/settings.js';