+++ packages/core/src/prompts/text.ts
@@ -1,5 +1,5 @@
 import color from 'picocolors';
-import Prompt, { type PromptOptions } from './prompt';
+import Prompt, { type PromptOptions } from './prompt.js';
 
 export interface TextOptions extends PromptOptions<TextPrompt> {
 	placeholder?: string;
@@ -7,7 +7,17 @@ export interface TextOptions extends PromptOptions<TextPrompt> {
 }
 
 export default class TextPrompt extends Prompt {
-	valueWithCursor = '';
+	get valueWithCursor() {
+		if (this.state === 'submit') {
+			return this.value;
+		}
+		if (this.cursor >= this.value.length) {
+			return `${this.value}█`;
+		}
+		const s1 = this.value.slice(0, this.cursor);
+		const [s2, ...s3] = this.value.slice(this.cursor);
+		return `${s1}${color.inverse(s2)}${s3.join('')}`;
+	}
 	get cursor() {
 		return this._cursor;
 	}
@@ -17,16 +27,6 @@ export default class TextPrompt extends Prompt {
 		this.on('finalize', () => {
 			if (!this.value) {
 				this.value = opts.defaultValue;
-			}
-			this.valueWithCursor = this.value;
-		});
-		this.on('value', () => {
-			if (this.cursor >= this.value.length) {
-				this.valueWithCursor = `${this.value}${color.inverse(color.hidden('_'))}`;
-			} else {
-				const s1 = this.value.slice(0, this.cursor);
-				const s2 = this.value.slice(this.cursor);
-				this.valueWithCursor = `${s1}${color.inverse(s2[0])}${s2.slice(1)}`;
 			}
 		});
 	}